backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pymongo==4.6.3
motor==3.3.2
pydantic==2.6.4
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator
python-jose[cryptography]==3.3.0
httpx==0.27.0
aiofiles==23.2.1
loguru==0.7.2

# Phase 3 — Audio Extraction
ffmpeg-python==0.2.0

# Phase 4 — Speech-to-Text
openai-whisper

# Phase 5 — Text Preprocessing
regex
nltk

# Phase 6 — AI Summarization
google-generativeai

# Phase 6 (optional, local BART) — large download ~1.6 GB, install manually:
# transformers==4.40.0
# torch

# Phase 7 — Quiz Generation
# google-generativeai
